requirements.txt
msal
requests
chromadb
python-dotenv